exercicio atualizado

+++ 05-exercicio-parte-1.rb
@@ -13,7 +13,7 @@ end
 
 
 print "Por favor, informe o valor do raio: "
-raio = gets.to_f
+raio = gets.to_i
 
 raio= calcular_diametro(raio)
